Work about Beta 0.1.1
The "Destroyer" disk has been slightly modified.
Slightly changed the colour of the sky in the biomes of the mod.
Set music for the biomes in the mod.
Filling of the "Savanna" (terracotta in height, acacia bushes (which are also in "Dry Savanna") and wild boars at the top of the height).
Slightly changed the chances of loot drop from the "Evoker".
The "Judas Fruit" has been completely redeveloped as a fruit of a "Baobab" tree.
Added a new damage type to the "Tumbleweed".